TREGO_PACKS/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-environ
pillow
django-cors-headers
gunicorn